Fixed Announcement so cursor does not jump to end of input when editing. Added confirmation message to delete button. Added alert to Save button so user knows something happened.

+++ frontend/components/ScheduleEditor.js
@@ -87,7 +87,9 @@ const ScheduleEditor = () => {
         <Button
           className="deleteButton"
           onClick={() => {
-            deleteScheduleFromDB()
+            if (confirm('You are about to delete a schedule.  Are you sure?')) {
+              deleteScheduleFromDB()
+            }
           }}
         >
           <FontAwesomeIcon icon="trash" />
